[refactor] Changes dashboard to use RouteHandler to show members
