+++ public/scripts/admin/dashboard.jsx
@@ -1,8 +1,13 @@
 // This page contains the React classes to display the dashboard
 var React = require('react');
 var Member = require('../shared/member.jsx');
-var Navigation = require('react-router').Navigation;
-var Link = require('react-router').Link;
+
+var Router = require('react-router');
+var RouteHandler = Router.RouteHandler;
+var Navigation = Router.Navigation;
+var Link = Router.Link;
+
+var AddForm = require('./subcomponents/AddForm.jsx');
 
 var Dashboard = React.createClass({
   componentDidMount: function() {
@@ -12,14 +17,12 @@ var Dashboard = React.createClass({
       success: function(data) {
         data = JSON.parse(data);
         var state = {};
-
-        this.setState({
-          orgName: data.name,
-          members: data.members || []
-        });
+        state.orgName = data.name || null;
+        state.members = data.members || [];
+        this.setState(state);
       }.bind(this),
       error: function(jqXHR, status, error) {
-        console.error('Dashboard req error:', jqXHR, status, error);
+        console.error('Dashboard req error:', status, error, jqXHR);
       }
     });
   },
@@ -34,9 +37,10 @@ var Dashboard = React.createClass({
       <div>
         <h2>{this.state.orgName}</h2>
         <h3>Dashboard</h3>
-        <Link to="add">Add User</Link>
+        <Link to="/dashboard/add">Add User</Link>
         <a href="/client">Launch Client</a>
         <Directory members={this.state.members} />
+        <RouteHandler />
       </div>
     );
   }
@@ -45,7 +49,7 @@ var Dashboard = React.createClass({
 var Directory = React.createClass({
   mixins: [Navigation],
   memberClick: function(self) {
-    console.log('redirect to edit for ', self.props.data.first_name);
+    console.log('redirect to edit for', self.props.data.first_name);
     // transition to member profile/edit member page
   },
   render: function() {
@@ -65,66 +69,6 @@ var Directory = React.createClass({
   }
 });
 
-var AddForm = React.createClass({
-  mixins: [Navigation],
-  handleSubmit: function(e) {
-    e.preventDefault();
-    var member = {
-      first_name: this.refs.firstName.getDOMNode().value.trim(),
-      last_name: this.refs.lastName.getDOMNode().value.trim(),
-      email: this.refs.email.getDOMNode().value.trim(),
-      phone: this.refs.phone.getDOMNode().value.trim()
-    };
-    $.ajax({
-      url: '/api/add',
-      method: 'POST',
-      data: member,
-      succss: function(data) {
-        console.log('User added');
-        // TODO: Add confirmation of user add
-        this.transitionTo('dashboard');
-      }.bind(this),
-      error: function(jqXHR, status, error) {
-        console.error(status, error);
-        this.transitionTo('dashboard');
-      }.bind(this),
-      timeout: 2000
-    });
-    // this.transitionTo('dashboard');
-  },
-  render: function() {
-    return (
-      <div className="container">
-        <form className="col-xs-8" onSubmit={this.handleSubmit}>
-          <h3>Add User</h3>
-          <div className="form-group">
-            <label>First Name</label>
-            <input type="text" className="form-control" ref="firstName" />
-          </div>
-          <div className="form-group">
-            <label>Last Name</label>
-            <input type="text" className="form-control" ref="lastName" />
-          </div>
-          <div className="form-group">
-            <label>Title</label>
-            <input type="text" className="form-control" ref="title" />
-          </div>
-          <div className="form-group">
-            <label>Email Address</label>
-            <input type="email" className="form-control" ref="email" />
-          </div>
-          <div className="form-group">
-            <label>Phone Number</label>
-            <input type="tel" className="form-control" ref="phone" />
-          </div>
-          <button type="submit" className="btn btn-default">Add User</button>
-        </form>
-      </div>
-    );
-  }
-});
-
 module.exports = {
-  Dashboard: Dashboard,
-  AddForm: AddForm
+  Dashboard: Dashboard
 };
